+++ crates/platform/src/api/addresses/address/total.rs
@@ -1,14 +1,15 @@
 use crate::addresses::{AddressInfo, AddressesPath};
-use crate::{BlockfrostError, api::ApiResult, server::state::AppState};
+use crate::{api::ApiResult, server::state::AppState};
 use axum::extract::{Path, State};
 use bf_api_provider::types::AddressesContentTotalResponse;
 
 pub async fn route(
     Path(address_path): Path<AddressesPath>,
-    State(app_state): State<AppState>,
+    State(state): State<AppState>,
 ) -> ApiResult<AddressesContentTotalResponse> {
     let AddressesPath { address, asset: _ } = address_path;
-    let _ = AddressInfo::from_address(&address, app_state.config.network.clone())?;
+    let address_info = AddressInfo::from_address(&address, state.config.network.clone())?;
+    let data_node = state.data_node()?;
 
-    Err(BlockfrostError::not_found())
+    data_node.addresses().total(&address_info.address).await
 }